refactor: image_file join 및 각 저장 후 Long  id 반환하도록 keyHolder 추가

+++ src/main/java/nextstep/courses/repository/EnrollmentRepository.java
@@ -3,7 +3,7 @@ package nextstep.courses.repository;
 import nextstep.courses.domain.Enrollment;
 
 public interface EnrollmentRepository {
-    int save(Enrollment enrollment);
+    Long save(Enrollment enrollment);
 
     Enrollment findById(Long id);
 }
