+++ src/main/java/nextstep/courses/repository/ImageFileRepository.java
@@ -3,7 +3,7 @@ package nextstep.courses.repository;
 import nextstep.courses.domain.ImageFile;
 
 public interface ImageFileRepository {
-    int save(ImageFile imageFile);
+    Long save(ImageFile imageFile);
 
     ImageFile findById(long id);
 }
