+++ src/main/java/nextstep/courses/repository/SessionRepository.java
@@ -4,6 +4,6 @@ package nextstep.courses.repository;
 import nextstep.courses.domain.Session;
 
 public interface SessionRepository {
-    int save(Session session);
+    Long save(Session session);
     Session findById(long id);
 }
